move Synchronized Reader namespace
